// frontend-admin/src/api/generated/model/tseHealthResponseDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface TseHealthResponseDto {
  consecutiveFailures?: number;
  /** @nullable */
  estimatedRecoveryTimeUtc?: string | null;
  /** @nullable */
  lastCheckUtc?: string | null;
  /** @nullable */
  lastErrorMessageSafe?: string | null;
  /** @nullable */
  lastSuccessfulPingUtc?: string | null;
  /** @nullable */
  nonFiscalPendingQueueCount?: number | null;
  /** @nullable */
  status?: string | null;
}